Move `SimilarityIndex` class to similarity module
The similarity module is for calculating the similarities
