+++ selector/tests/test_similarity.py
@@ -29,8 +29,13 @@ import pkg_resources
 import pytest
 from numpy.testing import assert_almost_equal, assert_equal, assert_raises
 
-from selector.methods.similarity import NSimilarity, SimilarityIndex
-from selector.similarity import modified_tanimoto, pairwise_similarity_bit, tanimoto
+from selector.methods.similarity import NSimilarity
+from selector.similarity import (
+    SimilarityIndex,
+    modified_tanimoto,
+    pairwise_similarity_bit,
+    tanimoto,
+)
 
 
 def test_pairwise_similarity_bit_raises():
